pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pykrxauth"
dynamic = ["version"]
description = "KRX data scraping"
readme = "README.md"
requires-python = ">=3.10" # 3.8/3.9는 이제 놓아주는 것을 추천드립니다.
license = {text = "MIT"}
authors = [{name = "Jonghun Yoo", email = "jonghun.yoo@outlook.com"}]
keywords = ["KRX", "stock", "bond", "financial", "data", "scraping"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "requests>=2.32.0",
    "pandas>=2.2.0,<3.0",
    "numpy>=1.24.0,<2.0",
    "deprecated>=1.2.14",
    "multipledispatch>=1.0.0",
    "matplotlib>=3.8.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Jin2ouS/pykrxauth"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.9.0", # 최신 3.14 대응을 위해 버전 상향
    "pre-commit>=3.5.0",
    "vcrpy>=4.1.0",
    "pytest-vcr>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["pykrxauth", "pykrxauth.*"]

[tool.setuptools.package-data]
pykrxauth = ["*.ttf"]

[tool.setuptools_scm]
write_to = "pykrxauth/_version.py"
local_scheme = "no-local-version"

# Ruff - Python linter/formatter (integrates Black, isort, Flake8)
[tool.ruff]
line-length = 88
target-version = "py310" # 최소 지원 버전에 맞춤

# 최소한의 exclude만 남김 (Ruff 기본값에 대부분 의존)
# Ignore setuptools-scm generated version file; it's autogenerated and may
# contain typing annotations that ruff flags.
exclude = ["build", "dist", "pykrxauth/_version.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403", "F405"] # wildcard imports in __init__.py
"tests/**/*.py" = ["S101"] # allow assert statements in tests
"pykrxauth/stock/stock_api.py" = ["F811"] # multipledispatch function overloading

[tool.ruff.format]
quote-style = "double"

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests/integration"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests (default)",
    "live: marks tests as live server checks (not run in CI)",
    "cassette: specify VCR cassette file for test",
]

# Coverage configuration (optional)
[tool.coverage.run]
source = ["pykrxauth"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
